requirements.txt
Django
gunicorn
beautifulsoup4
requests
djangorestframework
psycopg2-binary
daphne
dj_database_url
python-dotenv
whitenoise
pytest-django